Convert IPX and ENET timeouts to PIC time
The IPX server-connect timeout, IPX ping-reply timeout, and ENET
modem connect timeout all used GetTicks() / GetTicksSince() as the
deadline counter. That's host wall-clock time, which keeps ticking
across emulator pauses; on resume the timeout would false-fire as
if the entire pause duration had been spent waiting for the remote
peer. Converted to PIC_FullIndex() deltas (frozen with pause) so the
timeout measures emulated time only.

Real pre-existing bug, not just defensive bookkeeping for the new
pause flow.

+++ src/hardware/serialport/misc_util.h
@@ -133,8 +133,10 @@ private:
 	void updateState();
 
 #ifndef ENET_BLOCKING_CONNECT
-	int64_t connectStart = 0;
-	bool connecting      = false;
+	// PIC time (milliseconds since emulator start) so the connect
+	// timeout does not false-fire across an emulator pause.
+	double connectStart = 0.0;
+	bool connecting     = false;
 #endif
 	ENetHost* client                  = nullptr;
 	ENetPeer* peer                    = nullptr;
